+++ src/App.jsx
@@ -2,7 +2,9 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Header from "./components/Header";
 import Home from "./pages/Home";
+import Recipe from "./pages/Recipe";
 import './App.scss';
+import Surprise from "./pages/Surprise";
 
 function App() {
 
@@ -11,6 +13,8 @@ function App() {
       <Header />
       <Routes>
         <Route path="/" element={<Home />} />
+        <Route path="/recipe/:id" element={<Recipe />} />
+        <Route path="/recipe/surprise" element={<Surprise />} />
       </Routes>
     </Router>
   )