+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2024-12-21
+
+### Added
+- Search bar
+- Combined meals (by title+by ingredient)
+- Results grid
+- Surprise Me! button
+- Individual Recipe (placeholder)
+- Pagination
+
 ## [0.1.0] - 2024-12-21
 
 ### Added